+++ src/components/functional/Auth.tsx
@@ -1,13 +1,16 @@
+import type { VFC } from "react";
+
+import Loading from "@/components/common/Loading";
 import { useAuth } from "lib/auth";
 
 type Props = {
   children: JSX.Element;
 };
 
-const Auth = ({ children }: Props): JSX.Element => {
+const Auth: VFC<Props> = ({ children }) => {
   const isLoading = useAuth();
 
-  return isLoading ? <p>Loading...</p> : children;
+  return isLoading ? <Loading /> : children;
 };
 
 export default Auth;
