+++ src/pages/_app.tsx
@@ -1,7 +1,7 @@
 import type { AppProps } from "next/app";
 import { RecoilRoot } from "recoil";
 
-import { ThemeProvider } from "@/components/functional/Theme";
+import Theme from "@/components/functional/Theme";
 import Auth from "@/components/functional/Auth";
 import Layout from "@/components/common/Layout";
 import "@/styles/globals.css";
@@ -9,13 +9,13 @@ import "@/styles/globals.css";
 function MyApp({ Component, pageProps }: AppProps): JSX.Element {
   return (
     <RecoilRoot>
-      <ThemeProvider>
+      <Theme>
         <Auth>
           <Layout>
             <Component {...pageProps} />
           </Layout>
         </Auth>
-      </ThemeProvider>
+      </Theme>
     </RecoilRoot>
   );
 }
